refactor: split updates into state and nav, move with new namespacing
